+++ src/main/java/com/vault/demo/service/backstage/adxmn/selevicexmn.java
@@ -23,6 +23,6 @@ public interface selevicexmn {
     List<String> selectgetBytenid(int id);//根据标id查询出哪些用户
     List<Tender> slecttendermoney(int bid,int uid);//根据标和用户id查询金额
     Float seleUsermoney(int id);//根据id查询出用户余额
-    void updateuserMoney(Float money,int id);//根据用户id相加余额
+    void updateuserMoney(double money,int id);//根据用户id相加余额
     void addusermoney(Recharge recharge);//新增用户余额增加记录
 }
